requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# prompts
langchain_core


# Pydantic
Pydantic

# Document loaders
langchain_community
beautifulsoup4